Agrego LogIn y corrijo atributo User en Bot Telegram

+++ tacs-telegram-bot/src/main/java/org/tacsbot/handlers/LoginHandler.java
@@ -43,7 +43,7 @@ public class LoginHandler implements CommandsHandler {
                     } catch (JsonProcessingException e) {
                         throw new RuntimeException(e);
                     }
-                    bot.usersLoginMap.put(chatId,user.id.longValue());
+                    bot.usersLoginMap.put(chatId,user.id);
                     System.out.println(user.id);
                     System.out.println(bot.usersLoginMap.get(chatId));
                     bot.sendText(chatId,"Se ha logueado el usuario " + user.id.toString());
